refactor(ui): split into files
